Fix #33: Increase TEXT_BUFFER_SIZE to 275.
Inform7 is capable of overflowing the text (output) buffer by cramming
it with a maximum of 263 bytes.  This was discovered by pushing the
limits of the Inform7 compiler.  Inform6 won't do this, even if it's
told to print text far exceeding 263 bytes.

This overflow situation is not a security problem because the VM catches
it and aborts.  I am unsure if it's a good idea to truncate the attempt
to overfill the buffer and continue.