Modified the manpages in anticipation of the -W option
The -W option restricts Frotz to writing only in the specified
directory.  This is to prevent players from writing somewhere they
should not.  This is intended to be useful for front-ends or restricted
environments -- for instance a chatbot or telnetting to a certain TCP/IP
port to play a game.