Have resize_screen repaint window 0.
resize_screen decides which part of the window it wants to keep,
adjusts the window properties and calls a new front end function
os_repaint_window to repaint the required part.  Only window 0 currently
has this special handling - for the others it seems to suffice that
curses repaints the screen from the top left.  Resizing now
works pretty well in Beyond Zork as long as the width is not resized
and the height is enough for the upper window.

Also have resize_screen ask V6 games to redraw.  This used to be in
the front end but I can't see how any front end could do anything else
because the interpreter does not know what kind of window layout the
game wants.  So it may as well be in the common code.

#42: Handle terminal resizing