Adjust graphics font.
Set double height graphics font ascent to 14 and descent to 2
(was (12, 4)).  The new values match those of the built-in text font.
Set the y offset of all glyphs to -2 for both single and double
height graphics font, which now have the same descent
(yof was erroneously 0).  Beyond Zork now looks about right with the
built-in fonts.  It still looks awful with custom fonts because the
graphics font should be scaled to the size of the fixed width text
font.