Support for loading a scalable graphics font.
A graphics font can be specified with the "graphics" option
(Fonts section) of .sfrotzrc.  The font is assumed to have fixed width
equal to height and em size (no line gap).  It is scaled to match the
width and height of the fixed width text font (not usually equal,
so the graphics font gets stretched).  Any size specification in the
option is ignored.

Now someone just needs to draw that font.  I made a quick autotrace
in Fontforge but it is rather hideous.